requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.0
anthropic==0.40.0
openai==1.58.1
python-dotenv==1.0.1
pyyaml==6.0.2
aiofiles==24.1.0
python-multipart==0.0.9
jinja2==3.1.4
requests==2.32.3
aiohttp==3.9.0